Update Android-Components to 145.0
This version of android-components makes some changes to the bookmarks
API, methods that previously returned BookmarkNode or String, now return
a result and the actual value must be accessed with the getOrNull()
method. Minor fix was needed to BoomarksTest, the new folder is no longer
being selected by default when editing a bookmark.